only check suffix of instruction sequence for bad instructions
we know the prefix (i.e. RET) is correct, otherwise we wouldnt
consider it a gadget. this way we can check for other terminating
instructions (like RET) within the gadget even if those instructions
are the desired prefix. this is important because we dont define
a gadget as a sequence ending in RET; instead the user defines what the
ending should be (maybe a JMP %eax for example), but we still
dont want a gadget with a RET in the middle of it.