WIP: refactoring to allow for disassembler plugins
reliable and accurate disassembly for arbitrary
architectures is hard to come by, so sefi should
make it as easy as possible to add in support for
a new disassembler library for any architecture

a disassembler library will be encapsulated by
the Disassembler class and the Instr class. these
classes will implement a generic interface needed
by sefi to find instruction sequences/gadgets.