moved llvm.py -> sefi_llvm.py and distorm.py -> sefi_distorm.py
need to have a prefix so that an import like 'import llvm' will
actually import the llvm library instead of just importing the
sefi llvm file.

also made some random fixes to sefi_llvm.py. the most important one
is that i added a BadLLVMInstr class to represent instructions which
llvm couldnt decode at all (as opposed to decoding the instruction
but recognizing it as bad in some way or invalid even though it has
a valid encoding)